requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
jupyter>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
shap>=0.42.0
datasets>=2.14.0
huggingface-hub>=0.16.0

